Add support for choosing full vs qa, fix args

+++ refresh.sh
@@ -16,17 +16,17 @@
 # Argument defaults
 env_file=".env"
 
-while getopts ":f::s::h" opt; do
+while getopts ":h::e::c:" opt; do
   case $opt in
     h )
       echo "Usage:"
       echo "    refresh.sh -h                           Display this help message."
-      echo "    refresh.sh -f <environment file>        Specify an environment file to load from, defaults to .env."
+      echo "    refresh.sh -e <environment>             Specify an environment to update, either full or qa."
       echo "    refresh.sh -c <circle ci token>         Provide a CircleCI token."
       exit 0
       ;;
-    f )
-      env_file=$OPTARG
+    e )
+      env=$OPTARG
       ;;
     c )
       circle_ci_token=$OPTARG
@@ -47,12 +47,25 @@ echo "loaded COMMUNITY_LOGIN_URL=$COMMUNITY_LOGIN_URL"
 echo "loaded SALESFORCE_SECURITY_TOKEN=$SALESFORCE_SECURITY_TOKEN"
 echo "loaded E2E_SALESFORCE_PASSWORD=$E2E_SALESFORCE_PASSWORD"
 
-echo "Starting Heroku credential update for Webapp full"
+echo "Starting Heroku credential update for Partners $env"
 
-# Get all apps that are dahlia-lap-full apps.
-lap_full_apps=$(heroku apps --team=sfdigitalservices --json | jq '.[].name | select(test("dahlia-lap-full-*"))' )
+if [ $env == "full" ]; then
+  # Get all apps that are dahlia-web-full apps.
+  heroku_apps=$(heroku apps --team=sfdigitalservices --json | jq '.[].name | select(test("dahlia-lap-full-*"))')
+  heroku_apps=("dahlia-full" ${heroku_apps[@]})
+elif [ $env == "qa" ]; then
+  heroku_apps=('dahlia-lap-qa')
+else
+  echo "Error: environment must be full or qa"
+  exit 1
+fi
 
-for app in $lap_full_apps
+
+for app in ${heroku_apps[@]}
+  do
+    # Strip out double quotes from app names
+    app=$(echo "$app" | tr -d '"')
+    echo "Updating credentials for $app"
   do
     echo "Updating credentials for $app"
     heroku config:set SALESFORCE_CLIENT_SECRET=$SALESFORCE_CLIENT_SECRET --app $app
@@ -64,26 +77,27 @@ done
 
 echo "Heroku update complete"
 
+if [ $env == "full" ]; then
+  echo "Starting CircleCI credential update"
+  BASE_CIRCLECI_URL="https://circleci.com/api/v1.1/project/github/SFDigitalServices/sf-dahlia-lap/envvar"
 
-echo "Starting CircleCI credential update"
-BASE_CIRCLECI_URL="https://circleci.com/api/v1.1/project/github/SFDigitalServices/sf-dahlia-lap/envvar"
+  # Delete existing env vars
+  curl -X DELETE $BASE_CIRCLECI_URL/SALESFORCE_PASSWORD?circle-token=$circle_ci_token
+  curl -X DELETE $BASE_CIRCLECI_URL/SALESFORCE_SECURITY_TOKEN?circle-token=$circle_ci_token
+  curl -X DELETE $BASE_CIRCLECI_URL/SALESFORCE_CLIENT_SECRET?circle-token=$circle_ci_token
+  curl -X DELETE $BASE_CIRCLECI_URL/SALESFORCE_CLIENT_ID?circle-token=$circle_ci_token
+  curl -X DELETE $BASE_CIRCLECI_URL/SALESFORCE_INSTANCE_URL?circle-token=$circle_ci_token
+  curl -X DELETE $BASE_CIRCLECI_URL/COMMUNITY_LOGIN_URL?circle-token=$circle_ci_token
+  curl -X DELETE $BASE_CIRCLECI_URL/E2E_SALESFORCE_PASSWORD?circle-token=$circle_ci_token
 
-# Delete existing env vars
-curl -X DELETE $BASE_CIRCLECI_URL/SALESFORCE_PASSWORD?circle-token=$circle_ci_token
-curl -X DELETE $BASE_CIRCLECI_URL/SALESFORCE_SECURITY_TOKEN?circle-token=$circle_ci_token
-curl -X DELETE $BASE_CIRCLECI_URL/SALESFORCE_CLIENT_SECRET?circle-token=$circle_ci_token
-curl -X DELETE $BASE_CIRCLECI_URL/SALESFORCE_CLIENT_ID?circle-token=$circle_ci_token
-curl -X DELETE $BASE_CIRCLECI_URL/SALESFORCE_INSTANCE_URL?circle-token=$circle_ci_token
-curl -X DELETE $BASE_CIRCLECI_URL/COMMUNITY_LOGIN_URL?circle-token=$circle_ci_token
-curl -X DELETE $BASE_CIRCLECI_URL/E2E_SALESFORCE_PASSWORD?circle-token=$circle_ci_token
+  # Create new env vars
+  curl -X POST --header "Content-Type: application/json" -d '{"name": "SALESFORCE_PASSWORD", "value": "$SALESFORCE_PASSWORD"}' $BASE_CIRCLECI_URL?circle-token=$circle_ci_token
+  curl -X POST --header "Content-Type: application/json" -d '{"name": "SALESFORCE_SECURITY_TOKEN", "value": "$SALESFORCE_SECURITY_TOKEN"}' $BASE_CIRCLECI_URL?circle-token=$circle_ci_token
+  curl -X POST --header "Content-Type: application/json" -d '{"name": "SALESFORCE_CLIENT_SECRET", "value": "$SALESFORCE_CLIENT_SECRET"}' $BASE_CIRCLECI_URL?circle-token=$circle_ci_token
+  curl -X POST --header "Content-Type: application/json" -d '{"name": "SALESFORCE_CLIENT_ID", "value": "$SALESFORCE_CLIENT_ID"}' $BASE_CIRCLECI_URL?circle-token=$circle_ci_token
+  curl -X POST --header "Content-Type: application/json" -d '{"name": "SALESFORCE_INSTANCE_URL", "value": "$SALESFORCE_INSTANCE_URL"}' $BASE_CIRCLECI_URL?circle-token=$circle_ci_token
+  curl -X POST --header "Content-Type: application/json" -d '{"name": "COMMUNITY_LOGIN_URL", "value": "$COMMUNITY_LOGIN_URL"}' $BASE_CIRCLECI_URL?circle-token=$circle_ci_token
+  curl -X POST --header "Content-Type: application/json" -d '{"name": "E2E_SALESFORCE_PASSWORD", "value": "$E2E_SALESFORCE_PASSWORD"}' $BASE_CIRCLECI_URL?circle-token=$circle_ci_token
 
-# Create new env vars
-curl -X POST --header "Content-Type: application/json" -d '{"name": "SALESFORCE_PASSWORD", "value": "$SALESFORCE_PASSWORD"}' $BASE_CIRCLECI_URL?circle-token=$circle_ci_token
-curl -X POST --header "Content-Type: application/json" -d '{"name": "SALESFORCE_SECURITY_TOKEN", "value": "$SALESFORCE_SECURITY_TOKEN"}' $BASE_CIRCLECI_URL?circle-token=$circle_ci_token
-curl -X POST --header "Content-Type: application/json" -d '{"name": "SALESFORCE_CLIENT_SECRET", "value": "$SALESFORCE_CLIENT_SECRET"}' $BASE_CIRCLECI_URL?circle-token=$circle_ci_token
-curl -X POST --header "Content-Type: application/json" -d '{"name": "SALESFORCE_CLIENT_ID", "value": "$SALESFORCE_CLIENT_ID"}' $BASE_CIRCLECI_URL?circle-token=$circle_ci_token
-curl -X POST --header "Content-Type: application/json" -d '{"name": "SALESFORCE_INSTANCE_URL", "value": "$SALESFORCE_INSTANCE_URL"}' $BASE_CIRCLECI_URL?circle-token=$circle_ci_token
-curl -X POST --header "Content-Type: application/json" -d '{"name": "COMMUNITY_LOGIN_URL", "value": "$COMMUNITY_LOGIN_URL"}' $BASE_CIRCLECI_URL?circle-token=$circle_ci_token
-curl -X POST --header "Content-Type: application/json" -d '{"name": "E2E_SALESFORCE_PASSWORD", "value": "$E2E_SALESFORCE_PASSWORD"}' $BASE_CIRCLECI_URL?circle-token=$circle_ci_token
-
-echo "Credentials updated for CircleCI"
+  echo "Credentials updated for CircleCI"
+fi
